src/frontend/pyproject.toml
[project]
name = "mtl-finder-frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
]
